Move build-scripts/yaml-to-shorthand.py to ssg/_build_yaml.py
